+++ rwf-admin/src/controllers/index.rs
@@ -1,0 +1,11 @@
+use rwf::prelude::*;
+
+#[derive(Default)]
+pub struct Index;
+
+#[async_trait]
+impl Controller for Index {
+    async fn handle(&self, _request: &Request) -> Result<Response, Error> {
+        Ok(Response::new().redirect("jobs"))
+    }
+}
