+++ rwf/src/prelude.rs
@@ -16,4 +16,5 @@ pub use tokio;
 
 pub use macros::{context, crud, engine, render, rest, route};
 pub use rwf_macros as macros;
+pub use serde::{Deserialize, Serialize};
 pub use uuid::Uuid;
